Fix: Restructure - separate dev environment from output for importing
