+++ @blocksTS/folder/folder.d.ts
