--- @blocksTS/index.d.ts
@@ -1,2 +1,0 @@
-export * from './folder';
-export * from './prettify';
