+++ @blocksTS/prettify/prettify.d.ts
