+++ index.d.ts
@@ -1,0 +1,2 @@
+export * from './@blocksTS/folder/folder';
+export * from './@blocksTS/prettify/prettify';
