+++ index.js
@@ -14,5 +14,5 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
     for (var p in m) if (p !== "default" && !Object.prototype.hasOwnProperty.call(exports, p)) __createBinding(exports, m, p);
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-__exportStar(require("./folder"), exports);
-__exportStar(require("./prettify"), exports);
+__exportStar(require("./@blocksTS/folder/folder"), exports);
+__exportStar(require("./@blocksTS/prettify/prettify"), exports);
